Rename package "params.byname" to "bynameparams"
