+++ org.scala-ide.sdt.core.tests/src/org/scalaide/ui/internal/editor/decorators/bynameparams/CallByNameParamAtCreationPresenterTest.scala
@@ -1,4 +1,4 @@
-package org.scalaide.ui.internal.editor.decorators.params.byname
+package org.scalaide.ui.internal.editor.decorators.bynameparams
 
 import scala.annotation.migration
